requirements.txt
# main
Django==4.2.6
djangorestframework==3.14.0

# Security
django-environ==0.11.2
django-cors-headers==4.3.0

# For OpenAPI
drf-spectacular==0.26.5
drf-standardized-errors==0.12.5

# db
psycopg2-binary==2.9.9
django-filter==23.3

# celery
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

#google
oauth2client==4.1.3
google-api-python-client==2.108.0

